feat: 시그널 송신 API 개발 #BACK-365 Related to : #365

+++ backend/src/main/java/signiel/heartsigniel/model/meeting/dto/SingleSignalRequest.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class SingleSignalRequest {
     private int signalSequence;
-    private Long senderId;
-    private Long receiverId;
+    private int senderId;
+    private int receiverId;
 }
